BREAKING(cli/unstable): move `spinner` module to `unstable-spinner`
