+++ cli/deno.json
@@ -5,7 +5,7 @@
     ".": "./mod.ts",
     "./parse-args": "./parse_args.ts",
     "./prompt-secret": "./prompt_secret.ts",
-    "./spinner": "./spinner.ts",
+    "./unstable-spinner": "./unstable_spinner.ts",
     "./unicode-width": "./unicode_width.ts"
   }
 }
