+++ cli/mod.ts
@@ -17,5 +17,4 @@
 
 export * from "./parse_args.ts";
 export * from "./prompt_secret.ts";
-export * from "./spinner.ts";
 export * from "./unicode_width.ts";
