+++ cli/testdata/spinner_cases/custom_color_blue.ts
@@ -1,4 +1,4 @@
-import { Spinner } from "../../spinner.ts";
+import { Spinner } from "../../unstable_spinner.ts";
 
 const spinner = new Spinner({ color: "blue" });
