+++ cli/testdata/spinner_cases/custom_spinner.ts
@@ -1,4 +1,4 @@
-import { Spinner } from "../../spinner.ts";
+import { Spinner } from "../../unstable_spinner.ts";
 
 const spinner = new Spinner({
   spinner: ["0", "1", "2", "3", "4", "5", "6", "7", "8", "9"],
