+++ cli/unstable_spinner.ts
@@ -84,7 +84,7 @@ export interface SpinnerOptions {
  *
  * @example Usage
  * ```ts no-eval
- * import { Spinner } from "@std/cli/spinner";
+ * import { Spinner } from "@std/cli/unstable-spinner";
  *
  * const spinner = new Spinner({ message: "Loading...", color: "yellow" });
  * spinner.start();
@@ -104,7 +104,7 @@ export class Spinner {
    *
    * @example Usage
    * ```ts no-eval
-   * import { Spinner } from "@std/cli/spinner";
+   * import { Spinner } from "@std/cli/unstable-spinner";
    *
    * const spinner = new Spinner({ message: "Working..." });
    * spinner.start();
@@ -155,7 +155,7 @@ export class Spinner {
    *
    * @example Usage
    * ```ts no-eval
-   * import { Spinner } from "@std/cli/spinner";
+   * import { Spinner } from "@std/cli/unstable-spinner";
    *
    * const spinner = new Spinner({ message: "Loading...", color: "yellow" });
    * spinner.start();
@@ -175,7 +175,7 @@ export class Spinner {
    *
    * @example Usage
    * ```ts no-assert
-   * import { Spinner } from "@std/cli/spinner";
+   * import { Spinner } from "@std/cli/unstable-spinner";
    *
    * const spinner = new Spinner({ message: "Loading", color: "blue" });
    *
@@ -192,7 +192,7 @@ export class Spinner {
    *
    * @example Usage
    * ```ts no-eval
-   * import { Spinner } from "@std/cli/spinner";
+   * import { Spinner } from "@std/cli/unstable-spinner";
    *
    * const spinner = new Spinner({ message: "Loading..." });
    * spinner.start();
@@ -231,7 +231,7 @@ export class Spinner {
    *
    * @example Usage
    * ```ts no-eval
-   * import { Spinner } from "@std/cli/spinner";
+   * import { Spinner } from "@std/cli/unstable-spinner";
    *
    * const spinner = new Spinner({ message: "Loading..." });
    * spinner.start();
